CLARA-1.1.0/requirements.txt
# ============================================================
# CLARA - Python Dependencies
# Tested on Python 3.8+, PyTorch 2.0+, CUDA 11.8+
# ============================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & PEFT
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0

# Image Processing
pillow>=9.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.6.0

# Logging & Monitoring (Optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Development (Optional)
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
